new features, route bindings fix, cached routes
feature1: It will now detect which module is active based on route. If the target route is defined in X Module, X Module will be marked as active. For this reason now there is a function called activeBoot which will only be called in boot time if the module is active. Checking procedure to determine a module is active or not can be turned off using protected $checkActive = false; in provider.
feature2: loading factories from modules.
routes: are now loaded only if they are not already cached.
fix: added middleware:bindings to module routes to prevent disabling laravel's default route bindings.